requirements.txt
pip
git+git://github.com/Rapptz/discord.py.git#egg=discord.py[voice]
youtube_dl
imgurpython
python-dateutil
